Modified 'about' page title and content. Made Home Page to make it different from the index Made home page link exclusive to authorized users Removed _sidebar.blade.php (now using _navbar.blade.php) Removed getPokemonList method and page (they are unused) Deleted TrainerPokemonController and moved methods to MyPokemonController.
