+++ resources/views/partials/_navbar.blade.php
@@ -6,9 +6,11 @@
 
   <div class="collapse navbar-collapse" id="navbarsExampleDefault">
     <ul class="navbar-nav mr-auto">
+      @auth
       <li class="nav-item">
-        <a class="nav-link" href="{{ route('pages.index')}}">Home <span class="sr-only">(current)</span></a>
+        <a class="nav-link" href="{{ route('home')}}">Home <span class="sr-only">(current)</span></a>
       </li>
+      @endauth
       <li class="nav-item">
         <a class="nav-link" href="{{ route('pages.about')}}">About</a>
       </li>
